Rename sanitizer_hsa folder to hsa.
